// src/Rest/model/tblDiagnosticoSVPaciente.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import {TblDiagnosticoPaciente} from "./tblDiagnosticoPaciente";
import {TblSignosVitales} from "./tblSignosVitales";

export interface TblDiagnosticoSVPaciente {
  idDiagnosticoSVPaciente?: number;
  devEstado?: string;
  dsvDetalle?: string;
  tblDiagnosticoPaciente?: TblDiagnosticoPaciente
  tblSignosVitales?: TblSignosVitales
}